+++ src/query-builder-macros.php
@@ -1,6 +1,8 @@
 <?php
 
 use Illuminate\Database\Query\Builder as QueryBuilder;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Str;
 
 /**
  * Natural sorting.
@@ -25,3 +27,27 @@ QueryBuilder::macro(
         );
     }
 );
+
+// Add an Eloquent "whereLike" query builder macro
+QueryBuilder::macro('whereLike', function ($attributes, $searchTerm) {
+    $searchTerm = str_replace(' ', '%', $searchTerm);
+
+    $this->where(function (QueryBuilder $query) use ($attributes, $searchTerm) {
+        foreach (Arr::wrap($attributes) as $attribute) {
+            $query->when(Str::contains($attribute, '.'),
+                function (QueryBuilder $query) use ($attribute, $searchTerm) {
+                    list($relationName, $relationAttribute) = explode('.', $attribute);
+
+                    $query->orWhereHas($relationName, function (QueryBuilder $query) use ($relationAttribute, $searchTerm) {
+                        $query->where($relationAttribute, 'LIKE', "%{$searchTerm}%");
+                    });
+                },
+                function (QueryBuilder $query) use ($attribute, $searchTerm) {
+                    $query->orWhere($attribute, 'LIKE', "%{$searchTerm}%");
+                }
+            );
+        }
+    });
+
+    return $this;
+});