+++ changelog.md
@@ -1,5 +1,32 @@
-Changelog for Laravel Database Extension
-========================================
+Changelog
+=========
+
+2.1.1 (2019-03-20)
+------------------
+
+- Moving the "whereLike" macro to the right place
+- Register macro in the boot() method instead of the register() one
+
+2.1.0 (2019-03-20)
+------------------
+
+- Add an Eloquent "whereLike" query builder macro
+
+2.0.0 (2019-03-11)
+------------------
+
+- Drop support for Laravel < 5.8
+- Replace getForeignKey method call by getForeignKeyName
+
+1.2.0 (2019-03-07)
+------------------
+
+- Add support for Laravel 5.8
+
+1.1.0 (2018-09-11)
+------------------
+
+- add support for Laravel 5.7
 
 1.0.0 (2018-07-20)
 ------------------